增加 表 ENGINE 定义

+++ tools/generator/src/main/kotlin/top/bettercode/summer/tools/generator/ddl/MysqlToDDL.kt
@@ -8,30 +8,22 @@ object MysqlToDDL : ToDDL() {
     override val quoteMark: String = "`"
     override val commentPrefix: String = "#"
 
-    override fun toDDLUpdate(
-            module: String,
-            oldTables: List<Table>,
-            tables: List<Table>,
-            out: Writer,
-            extension: GeneratorExtension
-    ) {
+    override fun toDDLUpdate(module: String, oldTables: List<Table>, tables: List<Table>, out: Writer, extension: GeneratorExtension) {
         out.appendLine("$commentPrefix ${extension.datasources[module]!!.url.substringBefore("?")}")
         out.appendLine()
         if (tables != oldTables) {
-            val prefixTableName =
-                    if (extension.enable("include-schema")) {
-                        "$quote${extension.datasource(module).schema}$quote."
-                    } else {
-                        ""
-                    }
+            val prefixTableName = if (extension.enable("include-schema")) {
+                "$quote${extension.datasource(module).schema}$quote."
+            } else {
+                ""
+            }
             val tableNames = tables.map { it.tableName }
             val oldTableNames = oldTables.map { it.tableName }
-            if (extension.dropTablesWhenUpdate)
-                (oldTableNames - tableNames.toSet()).filter { "api_token" != it }.forEach {
-                    out.appendLine("$commentPrefix DROP $it")
-                    out.appendLine("DROP TABLE IF EXISTS $prefixTableName$quote$it$quote;")
-                    out.appendLine()
-                }
+            if (extension.dropTablesWhenUpdate) (oldTableNames - tableNames.toSet()).filter { "api_token" != it }.forEach {
+                out.appendLine("$commentPrefix DROP $it")
+                out.appendLine("DROP TABLE IF EXISTS $prefixTableName$quote$it$quote;")
+                out.appendLine()
+            }
             val newTableNames = tableNames - oldTableNames.toSet()
             tables.forEach { table ->
                 val tableName = table.tableName
@@ -41,15 +33,11 @@ object MysqlToDDL : ToDDL() {
                     val oldTable = oldTables.find { it.tableName == tableName }!!
                     if (oldTable != table) {
                         val lines = mutableListOf<String>()
-                        if (oldTable.remarks != table.remarks)
-                            lines.add(
-                                    "ALTER TABLE $prefixTableName$quote$tableName$quote COMMENT '${
-                                        table.remarks.replace(
-                                                "\\",
-                                                "\\\\"
-                                        )
-                                    }';"
-                            )
+                        if (oldTable.remarks != table.remarks) lines.add("ALTER TABLE $prefixTableName$quote$tableName$quote COMMENT '${
+                            table.remarks.replace("\\", "\\\\")
+                        }';")
+
+                        if (oldTable.engine != table.engine) lines.add("ALTER TABLE $prefixTableName$quote$tableName$quote ENGINE ${table.engine};")
 
                         val oldColumns = oldTable.columns
                         val columns = table.columns
@@ -60,14 +48,9 @@ object MysqlToDDL : ToDDL() {
                             val oldPrimaryKey = oldPrimaryKeys[0]
                             val primaryKey = primaryKeys[0]
                             if (oldPrimaryKey != primaryKey) {
-                                lines.add(
-                                        "ALTER TABLE $prefixTableName$quote$tableName$quote CHANGE $quote${oldPrimaryKey.columnName}$quote ${
-                                            columnDef(
-                                                    primaryKey,
-                                                    quote
-                                            )
-                                        } COMMENT '${primaryKey.remarks.replace("\\", "\\\\")}';"
-                                )
+                                lines.add("ALTER TABLE $prefixTableName$quote$tableName$quote CHANGE $quote${oldPrimaryKey.columnName}$quote ${
+                                    columnDef(primaryKey, quote)
+                                } COMMENT '${primaryKey.remarks.replace("\\", "\\\\")}';")
                                 oldColumns.remove(oldPrimaryKey)
                                 columns.remove(primaryKey)
                             }
@@ -81,41 +64,29 @@ object MysqlToDDL : ToDDL() {
                         val oldColumnNames = oldColumns.map { it.columnName }
                         val columnNames = columns.map { it.columnName }
                         val dropColumnNames = oldColumnNames - columnNames.toSet()
-                        if (extension.dropColumnsWhenUpdate)
-                            dropColumnNames.forEach {
-                                lines.add("ALTER TABLE $prefixTableName$quote$tableName$quote DROP COLUMN $quote$it$quote;")
-                            }
+                        if (extension.dropColumnsWhenUpdate) dropColumnNames.forEach {
+                            lines.add("ALTER TABLE $prefixTableName$quote$tableName$quote DROP COLUMN $quote$it$quote;")
+                        }
                         dropFk(prefixTableName, oldColumns, dropColumnNames, lines, tableName)
                         val newColumnNames = columnNames - oldColumnNames.toSet()
                         columns.forEach { column ->
                             val columnName = column.columnName
                             if (newColumnNames.contains(columnName)) {
-                                lines.add(
-                                        "ALTER TABLE $prefixTableName$quote$tableName$quote ADD COLUMN ${
-                                            columnDef(
-                                                    column,
-                                                    quote
-                                            )
-                                        } COMMENT '${column.remarks.replace("\\", "\\\\")}';"
-                                )
+                                lines.add("ALTER TABLE $prefixTableName$quote$tableName$quote ADD COLUMN ${
+                                    columnDef(column, quote)
+                                } COMMENT '${column.remarks.replace("\\", "\\\\")}';")
                                 addFk(prefixTableName, column, lines, tableName, columnName)
                             } else {
                                 val oldColumn = oldColumns.find { it.columnName == columnName }!!
                                 if (column != oldColumn) {
-                                    lines.add(
-                                            "ALTER TABLE $prefixTableName$quote$tableName$quote MODIFY ${
-                                                columnDef(
-                                                        column,
-                                                        quote
-                                                )
-                                            } COMMENT '${column.remarks.replace("\\", "\\\\")}';"
-                                    )
+                                    lines.add("ALTER TABLE $prefixTableName$quote$tableName$quote MODIFY ${
+                                        columnDef(column, quote)
+                                    } COMMENT '${column.remarks.replace("\\", "\\\\")}';")
                                     updateFk(prefixTableName, column, oldColumn, lines, tableName)
                                 }
                             }
                         }
-                        if (extension.datasources[module]?.queryIndex == true)
-                            updateIndexes(prefixTableName, oldTable, table, lines, dropColumnNames)
+                        if (extension.datasources[module]?.queryIndex == true) updateIndexes(prefixTableName, oldTable, table, lines, dropColumnNames)
                         if (lines.isNotEmpty()) {
                             out.appendLine("$commentPrefix $tableName")
                             lines.forEach { out.appendLine(it) }
@@ -128,51 +99,29 @@ object MysqlToDDL : ToDDL() {
         }
     }
 
-    override fun dropFkStatement(
-            prefixTableName: String,
-            tableName: String,
-            columnName: String
-    ): String =
-            "ALTER TABLE $prefixTableName$quote$tableName$quote DROP FOREIGN KEY ${
-                foreignKeyName(
-                        tableName,
-                        columnName
-                )
-            };"
+    override fun dropFkStatement(prefixTableName: String, tableName: String, columnName: String): String = "ALTER TABLE $prefixTableName$quote$tableName$quote DROP FOREIGN KEY ${
+        foreignKeyName(tableName, columnName)
+    };"
 
     override fun appendTable(prefixTableName: String, table: Table, pw: Writer) {
         val tableName = table.tableName
         pw.appendLine("$commentPrefix $tableName")
-        if (table.ext.dropTablesWhenUpdate)
-            pw.appendLine("DROP TABLE IF EXISTS $prefixTableName$quote$tableName$quote;")
+        if (table.ext.dropTablesWhenUpdate) pw.appendLine("DROP TABLE IF EXISTS $prefixTableName$quote$tableName$quote;")
         pw.appendLine("CREATE TABLE $prefixTableName$quote$tableName$quote (")
         val hasPrimary = table.primaryKeyNames.isNotEmpty()
         val lastIndex = table.columns.size - 1
         table.columns.forEachIndexed { index, column ->
-            pw.appendLine(
-                    "  ${
-                        columnDef(
-                                column,
-                                quote
-                        )
-                    } COMMENT '${
-                        column.remarks.replace(
-                                "\\",
-                                "\\\\"
-                        )
-                    }'${if (index < lastIndex || hasPrimary) "," else ""}"
-            )
+            pw.appendLine("  ${
+                columnDef(column, quote)
+            } COMMENT '${
+                column.remarks.replace("\\", "\\\\")
+            }'${if (index < lastIndex || hasPrimary) "," else ""}")
         }
 
         appendKeys(table, hasPrimary, pw, quote, tableName, useForeignKey)
-        pw.appendLine(
-                ")${if (table.physicalOptions.isNotBlank()) " ${table.physicalOptions}" else ""} COMMENT = '${
-                    table.remarks.replace(
-                            "\\",
-                            "\\\\"
-                    )
-                }';"
-        )
+        pw.appendLine(")${if (table.physicalOptions.isNotBlank()) " ${table.physicalOptions}" else ""} COMMENT = '${
+            table.remarks.replace("\\", "\\\\")
+        }'${if (table.engine.isNotBlank()) " ENGINE = ${table.engine};" else ""};")
 
         appendIndexes(prefixTableName, table, pw, quote)
